Store taints in node annotation
Storing taints in annotation helps to identify if the taint set on
node is owned by NFD or not. When user deletes the taint entry from
NodeFeatureRule CR, NFD will remove the taint from the node. But
to avoid accidental deletion of taints not owned by the NFD, it
needs to know the owner. Keeping track of NFD set taints in the
annotation can be used during the filtering of the owner.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@intel.com>