e2e: re-write pre-submit test to use minikube as an ephemeral cluster
Using minikube with Docker being the vm-driver simplifies our
pre-submit e2e tests with image registry handling. One of the problem
with using kind and local registry container is that we need to have
mounted the same port (5000) to all the kind clusters that might be
created at the same time. And that's technically impossible, because
we can't mount the same port to all the clusters. Minikube on the
other hand, allows pushing directly to the in-cluster Docker daemon.
It comes with its registry and we don't have to take care of it ourselves.
Every minikube cluster will have it is own registry that will be created
or destroyed as the cluster itself.

Signed-off-by: Muyassarov, Feruzjon <feruzjon.muyassarov@intel.com>